Refactor `trigger` module
