+++ src/trigger/mod.rs
@@ -7,10 +7,13 @@ use tokio_util::sync::CancellationToken;
 use tracing::{error, info};
 
 use crate::{
-    error::AppError, events::BranchUpdateEvent, model::Subscriber, trigger::jwt::generate_gh_jwt,
+    error::AppError,
+    events::BranchUpdateEvent,
+    model::Subscriber,
+    trigger::auth::{generate_gh_jwt, request_iat},
 };
 
-mod jwt;
+mod auth;
 
 pub fn start_trigger_engine(
     pool: SqlitePool,
@@ -97,41 +100,6 @@ async fn notify_subscriber(
     let iat = request_iat(http_client, jwt, &sub).await?;
     send_repository_dispatch(http_client, &iat, event, &sub).await?;
     Ok(())
-}
-
-async fn request_iat(
-    http_client: &Client,
-    jwt: &str,
-    sub: &Subscriber,
-) -> Result<String, AppError> {
-    #[derive(serde::Deserialize)]
-    struct IatResponse {
-        token: String,
-    }
-
-    let api_url = format!(
-        "https://api.github.com/app/installations/{}/access_tokens",
-        sub.gh_app_installation_id
-    );
-    let response = http_client
-        .post(&api_url)
-        .bearer_auth(jwt)
-        .header("Accept", "application/vnd.github+json")
-        .header("X-GitHub-Api-Version", "2026-03-10")
-        .send()
-        .await?;
-
-    if response.status().is_success() {
-        let response_json = response.json::<IatResponse>().await?;
-        info!("IAT received for subscriber {}", sub.target_repo);
-        Ok(response_json.token)
-    } else {
-        Err(AppError::Response(format!(
-            "Unexpected status {}: {}",
-            response.status(),
-            response.text().await?
-        )))
-    }
 }
 
 async fn send_repository_dispatch(
